[APG-454] Check if staff details exisit in db before fetching

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/client/nomisUserRoleManagementApi/model/StaffDetailResponse.kt
@@ -2,11 +2,10 @@ package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.client.nomisUs
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
-import java.math.BigInteger
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class StaffDetailResponse(
-  val staffId: BigInteger,
+  val staffId: Int,
   val firstName: String,
   val lastName: String,
   val status: String,
