+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/entity/create/ReferralEntity.kt
@@ -10,7 +10,6 @@ import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
 import jakarta.persistence.Version
 import org.hibernate.annotations.SQLRestriction
-import java.math.BigInteger
 import java.time.LocalDateTime
 import java.util.UUID
 
@@ -50,7 +49,7 @@ data class ReferralEntity(
   var deleted: Boolean = false,
 
   @Column(name = "primary_pom_staff_id")
-  var primaryPomStaffId: BigInteger? = null,
+  var primaryPomStaffId: Int? = null,
   @Column(name = "secondary_pom_staff_id")
-  var secondaryPomStaffId: BigInteger? = null,
+  var secondaryPomStaffId: Int? = null,
 )
