+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/repository/StaffRepository.kt
@@ -3,11 +3,10 @@ package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.reposit
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.entity.create.StaffEntity
-import java.math.BigInteger
 import java.util.UUID
 
 @Repository
 interface StaffRepository : JpaRepository<StaffEntity, UUID> {
 
-  fun findByStaffId(staffId: BigInteger): StaffEntity?
+  fun findByStaffId(staffId: Int): StaffEntity?
 }
