+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/restapi/model/StaffDetail.kt
@@ -1,10 +1,9 @@
 package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.model
 
 import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.entity.create.AccountType
-import java.math.BigInteger
 
 data class StaffDetail(
-  val staffId: BigInteger,
+  val staffId: Int,
   val firstName: String,
   val lastName: String,
   val primaryEmail: String,
